remove unneeded code/tests
