+++ internal/controller/mapserver/deployment_test.go
@@ -3,15 +3,10 @@ package mapserver
 import (
 	"testing"
 
-	"github.com/pdok/mapserver-operator/internal/controller/constants"
-
 	"github.com/pdok/mapserver-operator/api/v2beta1"
 	pdoknlv3 "github.com/pdok/mapserver-operator/api/v3"
-	"github.com/pdok/mapserver-operator/internal/controller/types"
-	smoothoperatorutils "github.com/pdok/smooth-operator/pkg/util"
 	"github.com/stretchr/testify/assert"
 	corev1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/api/resource"
 	"sigs.k8s.io/yaml"
 
 	_ "embed"
@@ -20,51 +15,14 @@ import (
 //go:embed test_data/expected_volumemounts.yaml
 var expectedVolumeMountsYaml []byte
 
-func TestGetVolumeMountsForDeployment(t *testing.T) {
-	var wfs = getV3()
+func TestGetVolumeMounts(t *testing.T) {
 	pdoknlv3.SetHost("https://service.pdok.nl")
-	result := GetVolumeMountsForDeployment(wfs)
+	result := getVolumeMounts(false)
 
 	var expectedVolumeMounts struct{ VolumeMounts []corev1.VolumeMount }
 	err := yaml.Unmarshal(expectedVolumeMountsYaml, &expectedVolumeMounts)
 	assert.NoError(t, err)
 	assert.Equal(t, expectedVolumeMounts.VolumeMounts, result)
-}
-
-//go:embed test_data/expected_envvars.yaml
-var expectedEnvVarsYaml []byte
-
-func TestGetEnvVarsForDeployment(t *testing.T) {
-	var wfs = getV3()
-	pdoknlv3.SetHost("https://service.pdok.nl")
-	result := GetEnvVarsForDeployment(wfs, "blobs-secret")
-	var expectedEnvVars struct{ EnvVars []corev1.EnvVar }
-	err := yaml.Unmarshal(expectedEnvVarsYaml, &expectedEnvVars)
-	assert.NoError(t, err)
-	assert.Equal(t, expectedEnvVars.EnvVars, result)
-}
-
-func TestGetResourcesForDeployment(t *testing.T) {
-	var wfs = getV3()
-	pdoknlv3.SetHost("https://service.pdok.nl")
-	result := GetResourcesForDeployment(wfs)
-
-	expectedLimits := corev1.ResourceList{}
-	expectedRequest := corev1.ResourceList{}
-
-	expectedLimits[corev1.ResourceMemory] = resource.MustParse("800M")
-	expectedLimits[corev1.ResourceEphemeralStorage] = resource.MustParse("505Mi")
-
-	expectedRequest[corev1.ResourceCPU] = resource.MustParse("0.15")
-	expectedRequest[corev1.ResourceEphemeralStorage] = resource.MustParse("255Mi")
-
-	var expected = corev1.ResourceRequirements{
-		Limits:   expectedLimits,
-		Requests: expectedRequest,
-		Claims:   nil,
-	}
-
-	assert.Equal(t, expected, result)
 }
 
 //go:embed test_data/expected_livenessprobe.yaml
@@ -79,7 +37,7 @@ var expectedStartupProbe []byte
 func TestGetProbesForDeployment(t *testing.T) {
 	var wfs = getV3()
 	pdoknlv3.SetHost("https://service.pdok.nl")
-	livenessResult, readinessResult, startupResult, err := GetProbesForDeployment(wfs)
+	livenessResult, readinessResult, startupResult, err := getProbes(wfs)
 	assert.NoError(t, err)
 
 	var expectedLiveness corev1.Probe
@@ -94,33 +52,6 @@ func TestGetProbesForDeployment(t *testing.T) {
 	assert.Equal(t, &expectedLiveness, livenessResult)
 	assert.Equal(t, &expectedReadiness, readinessResult)
 	assert.Equal(t, &expectedStartup, startupResult)
-}
-
-func TestGetVolumesForDeployment(t *testing.T) {
-	var wfs = getV3()
-	wfs.Spec.Options.PrefetchData = false
-	pdoknlv3.SetHost("https://service.pdok.nl")
-
-	hashedConfigMapNames := types.HashedConfigMapNames{
-		Mapserver:             "rws-nwbwegen-v1-0-wfs-mapserver-bb59c7f4f4",
-		InitScripts:           "2",
-		MapfileGenerator:      "rws-nwbwegen-v1-0-wfs-mapfile-generator-bbbtd999dh",
-		CapabilitiesGenerator: "rws-nwbwegen-v1-0-wfs-capabilities-generator-6m4mfkgb5d",
-		OgcWebserviceProxy:    "3",
-		LegendGenerator:       "4",
-		FeatureInfoGenerator:  "5",
-	}
-	result := GetVolumesForDeployment(wfs, hashedConfigMapNames)
-
-	expected := []corev1.Volume{
-		{Name: constants.BaseVolumeName, VolumeSource: corev1.VolumeSource{EmptyDir: &corev1.EmptyDirVolumeSource{}}},
-		{Name: constants.DataVolumeName, VolumeSource: corev1.VolumeSource{EmptyDir: &corev1.EmptyDirVolumeSource{}}},
-		{Name: constants.MapserverName, VolumeSource: corev1.VolumeSource{ConfigMap: &corev1.ConfigMapVolumeSource{LocalObjectReference: corev1.LocalObjectReference{Name: "rws-nwbwegen-v1-0-wfs-mapserver-bb59c7f4f4"}, DefaultMode: smoothoperatorutils.Pointer(int32(420))}}},
-		{Name: constants.ConfigMapCapabilitiesGeneratorVolumeName, VolumeSource: corev1.VolumeSource{ConfigMap: &corev1.ConfigMapVolumeSource{LocalObjectReference: corev1.LocalObjectReference{Name: "rws-nwbwegen-v1-0-wfs-capabilities-generator-6m4mfkgb5d"}, DefaultMode: smoothoperatorutils.Pointer(int32(420))}}},
-		{Name: constants.ConfigMapMapfileGeneratorVolumeName, VolumeSource: corev1.VolumeSource{ConfigMap: &corev1.ConfigMapVolumeSource{LocalObjectReference: corev1.LocalObjectReference{Name: "rws-nwbwegen-v1-0-wfs-mapfile-generator-bbbtd999dh"}, DefaultMode: smoothoperatorutils.Pointer(int32(420))}}},
-	}
-
-	assert.Equal(t, expected, result)
 }
 
 //go:embed test_data/v2_input.yaml
