--- internal/controller/mapserver/test_data/expected_envvars.yaml
@@ -1,12 +1,0 @@
-envVars:
-- name: SERVICE_TYPE
-  value: WFS
-- name: MAPSERVER_CONFIG_FILE
-  value: /srv/mapserver/config/default_mapserver.conf
-- name: MS_MAPFILE
-  value: /srv/data/config/mapfile/service.map
-- name: AZURE_STORAGE_CONNECTION_STRING
-  valueFrom:
-    secretKeyRef:
-      key: AZURE_STORAGE_CONNECTION_STRING
-      name: blobs-secret
